ascii: AsciiGallery
Port components/ascii/ascii-gallery.tsx → Svelte 5.

- AsciiGallery.svelte: stacked canvases with CSS opacity crossfade,
  cycling timer, and ResizeObserver all managed in onMount
- Uses SvelteSet instead of plain Set (autofixer fix)
- bind:this array pattern for indexed canvas refs (advisory only,
  kept as-is since indexed access is required by bake/show logic)
- Autofixer: 0 issues, 2 advisory bind:this suggestions (not applied)
